fix: update partition sizes for Windows 2024 (GE) support
Germanium introduces updated logic to compute the minimum system requirement for Minimum Disk size, which no longer verify the size of the disk, but the size of the Windows OS partition. As a result, update all sizes to ensure a minimum of 64GB for the Windows Primary Data Partition (where the OS resides). Also update the EFIESP sizes to be 260MB maximum.